split copypastegame.ts into copygame.ts and pastegame.ts
