+++ translation/changes.json
@@ -1,6 +1,6 @@
 {
 	"65": {
-		"note": "Deleted slidelimit_not_number line 397.",
+		"note": "Deleted slidelimit_not_number line 397. Changed invalid_wincon_white and invalid_wincon_black on line 389 to invalid_wincon.",
 		"changes": []
 	},
 	"64": {
